requirements.txt
# Dependências do OLXScrapper.py
selenium==4.15.2
webdriver-manager==4.0.1
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.5.2
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6 